+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `set_trigger` method
+- Reversed ranges (they already worked but that was an undocumented feature, now
+  it's part of the API)
 
 ## [0.3.0] - 2022-02-10